akkor_et_al_2026/requirements.txt
apturl==0.5.2
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
click==8.1.8
colorama==0.4.3
command-not-found==0.3
contourpy==1.3.0
cplex==22.1.0.0
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==1.16.0
docplex==2.23.221
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
flexcache==0.3
flexparser==0.4
fonttools==4.58.4
future==0.18.2
gssapi==1.6.1
gurobipy==9.5.1
html5lib==1.0.1
httplib2==0.14.0
idaes-pse==2.8.0
idna==2.8
importlib-resources==6.5.2
ipaclient==4.8.6
ipalib==4.8.6
ipaplatform==4.8.6
ipapython==4.8.6
Jinja2==2.10.1
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.9.4
monotonic==1.5
mpmath==1.3.0
netaddr==0.7.19
netifaces==0.10.4
networkx==3.2.1
numpy==2.0.2
oauthlib==3.1.0
olefile==0.46
packaging==25.0
pandas==2.3.0
paramiko==2.6.0
pexpect==4.6.0
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.8
ply==3.11
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.19
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyomo==6.9.1
pyparsing==3.2.3
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-augeas==0.5.0
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-ldap==3.2.0
python-yubico==1.3.3
pytz==2025.2
pyusb==1.0.2
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.13.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.5
ssh-import-id==5.10
sympy==1.14.0
systemd-python==234
typing-extensions==4.14.0
tzdata==2025.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
zipp==3.23.0
